fix: approval flow delete

+++ backend/crm/src/main/java/cn/cordys/crm/approval/service/ApprovalInstanceService.java
@@ -340,14 +340,18 @@ public class ApprovalInstanceService {
 		List<String> instanceIds = instances.stream().map(ApprovalInstance::getId).toList();
 		List<ApprovalTask> approvalTasks = approvalTaskMapper.selectListByLambda(new LambdaQueryWrapper<ApprovalTask>().in(ApprovalTask::getInstanceId, instanceIds));
 		List<String> taskIds = approvalTasks.stream().map(ApprovalTask::getId).toList();
-		approvalTaskMapper.deleteByIds(taskIds);
-		approvalAddSignTaskMapper.deleteByLambda(new LambdaQueryWrapper<ApprovalAddSignTask>().in(ApprovalAddSignTask::getTaskId, taskIds));
-		approvalReturnBackRecordMapper.deleteByLambda(new LambdaQueryWrapper<ApprovalReturnBackRecord>().in(ApprovalReturnBackRecord::getTaskId, taskIds));
-		approvalRecordMapper.deleteByLambda(new LambdaQueryWrapper<ApprovalRecord>().in(ApprovalRecord::getInstanceId, instanceIds));
-		List<ApprovalInstanceAttachment> instanceAttachments = approvalInstanceAttachmentMapper.selectListByLambda(new LambdaQueryWrapper<ApprovalInstanceAttachment>().in(ApprovalInstanceAttachment::getInstanceId, instanceIds));
-		approvalInstanceAttachmentMapper.deleteByIds(instanceAttachments.stream().map(ApprovalInstanceAttachment::getId).toList());
-		List<String> attachmentIds = instanceAttachments.stream().map(ApprovalInstanceAttachment::getAttachmentId).toList();
-		attachmentIds.forEach(id -> attachmentService.delete(id));
-		attachmentMapper.deleteByIds(attachmentIds);
+		if (CollectionUtils.isNotEmpty(taskIds)) {
+			approvalTaskMapper.deleteByIds(taskIds);
+			approvalAddSignTaskMapper.deleteByLambda(new LambdaQueryWrapper<ApprovalAddSignTask>().in(ApprovalAddSignTask::getTaskId, taskIds));
+			approvalReturnBackRecordMapper.deleteByLambda(new LambdaQueryWrapper<ApprovalReturnBackRecord>().in(ApprovalReturnBackRecord::getTaskId, taskIds));
+		}
+		if (CollectionUtils.isNotEmpty(taskIds)) {
+			approvalRecordMapper.deleteByLambda(new LambdaQueryWrapper<ApprovalRecord>().in(ApprovalRecord::getInstanceId, instanceIds));
+			List<ApprovalInstanceAttachment> instanceAttachments = approvalInstanceAttachmentMapper.selectListByLambda(new LambdaQueryWrapper<ApprovalInstanceAttachment>().in(ApprovalInstanceAttachment::getInstanceId, instanceIds));
+			approvalInstanceAttachmentMapper.deleteByIds(instanceAttachments.stream().map(ApprovalInstanceAttachment::getId).toList());
+			List<String> attachmentIds = instanceAttachments.stream().map(ApprovalInstanceAttachment::getAttachmentId).toList();
+			attachmentIds.forEach(id -> attachmentService.delete(id));
+			attachmentMapper.deleteByIds(attachmentIds);
+		}
 	}
 }
